Added authentication to create and delete endpoints

+++ part4/blog/controllers/blogs.js
@@ -2,6 +2,7 @@ import logger from "../utils/logger.js";
 import express from "express";
 import Blog from "../models/blog.js";
 import User from "../models/user.js";
+import jwt from "jsonwebtoken";
 
 const blogsRouter = express.Router();
 
@@ -18,38 +19,45 @@ blogsRouter.get("/", async (request, response) => {
 });
 
 blogsRouter.post("/", async (request, response) => {
-  try {
-    if (!request.body.likes) request.body.likes = 0;
-
-    if (!request.body.title || !request.body.url)
-      return response.status(400).json({
-        error: "title or url missing",
-      });
-
-    const users = await User.find({});
-    const user = users[0];
-
-    const blog = new Blog({ ...request.body, user: user._id });
-    const savedBlog = await blog.save();
-
-    user.blogs = user.blogs.concat(savedBlog._id);
-    await user.save();
-
-    return response.status(201).json(savedBlog);
-  } catch (err) {
-    logger.error(err);
-    return response.status(500).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
   }
+  const user = await User.findById(decodedToken.id);
+
+  if (!request.body.likes) request.body.likes = 0;
+
+  if (!request.body.title || !request.body.url)
+    return response.status(400).json({
+      error: "title or url missing",
+    });
+
+  const blog = new Blog({ ...request.body, user: user._id });
+  const savedBlog = await blog.save();
+
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  return response.status(201).json(savedBlog);
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  try {
-    const result = await Blog.findByIdAndDelete(request.params.id);
-    return response.status(201).json(result);
-  } catch (err) {
-    logger.error(err);
-    return response.status(500).json({ error: err.message });
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
   }
+
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (!blogToDelete) return response.status(204).end();
+
+  if (blogToDelete.user.toString() !== decodedToken.id.toString())
+    return response
+      .status(401)
+      .json({ error: "blog can only be deleted by author" });
+
+  await Blog.findByIdAndDelete(request.params.id);
+  return response.status(204).end();
 });
 
 blogsRouter.put("/:id", async (request, response) => {
@@ -61,19 +69,14 @@ blogsRouter.put("/:id", async (request, response) => {
       error: "title or url missing",
     });
 
-  try {
-    const result = await Blog.findByIdAndUpdate(
-      request.params.id,
-      {
-        likes: request.body.likes,
-      },
-      { returnDocument: "after", runValidators: true },
-    );
-    return response.status(201).json(result);
-  } catch (err) {
-    logger.error(err);
-    return response.status(500).end();
-  }
+  const result = await Blog.findByIdAndUpdate(
+    request.params.id,
+    {
+      likes: request.body.likes,
+    },
+    { returnDocument: "after", runValidators: true },
+  );
+  return response.status(201).json(result);
 });
 
 export default blogsRouter;
